# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tools/lesson-planner": {
      "post": {
        "summary": "Generate Lesson Plan",
        "operationId": "generate_lesson_plan_api_tools_lesson_planner_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/visual-generator": {
      "post": {
        "summary": "Generate Visuals",
        "operationId": "generate_visuals_api_tools_visual_generator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisualGeneratorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeneratedVisual"
                  },
                  "type": "array",
                  "title": "Response Generate Visuals Api Tools Visual Generator Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/hyper-local": {
      "post": {
        "summary": "Generate Hyper Local Content",
        "operationId": "generate_hyper_local_content_api_tools_hyper_local_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HyperLocalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HyperLocalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/story-generator": {
      "post": {
        "summary": "Generate Story",
        "operationId": "generate_story_api_tools_story_generator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryGeneratorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/quiz-generator": {
      "post": {
        "summary": "Generate Quiz",
        "operationId": "generate_quiz_api_tools_quiz_generator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizGeneratorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/rubric-generator": {
      "post": {
        "summary": "Generate Rubric",
        "operationId": "generate_rubric_api_tools_rubric_generator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RubricGeneratorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/simulation-generator": {
      "post": {
        "summary": "Generate Simulation",
        "operationId": "generate_simulation_api_tools_simulation_generator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationGeneratorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/paper-digitizer": {
      "post": {
        "summary": "Digitize Paper",
        "operationId": "digitize_paper_api_tools_paper_digitizer_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_digitize_paper_api_tools_paper_digitizer_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDigitizerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_digitize_paper_api_tools_paper_digitizer_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_digitize_paper_api_tools_paper_digitizer_post"
      },
...
      "CurriculumMappingItem": {
        "properties": {
          "standard": {
            "type": "string",
            "title": "Standard"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "standard",
          "description"
        ],
        "title": "CurriculumMappingItem"
      },
      "GeneratedVisual": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "style": {
            "type": "string",
            "title": "Style"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "style"
        ],
        "title": "GeneratedVisual"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HyperLocalRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "culturalContext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Culturalcontext"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "grade",
          "location",
          "language"
        ],
        "title": "HyperLocalRequest"
      },
      "HyperLocalResponse": {
        "properties": {
          "region": {
            "type": "string",
            "title": "Region"
          },
          "culturalContext": {
            "type": "string",
            "title": "Culturalcontext"
          },
          "content": {
            "items": {
              "$ref": "#/components/schemas/LocalizedContentItem"
            },
            "type": "array",
            "title": "Content"
          },
          "curriculumMapping": {
            "items": {
              "$ref": "#/components/schemas/CurriculumMappingItem"
            },
            "type": "array",
            "title": "Curriculummapping"
          }
        },
        "type": "object",
        "required": [
          "region",
          "culturalContext",
          "content",
          "curriculumMapping"
        ],
        "title": "HyperLocalResponse"
      },
      "LessonPlanRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "bookId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bookid"
          },
          "chapterIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapterids"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "grade",
          "duration"
        ],
        "title": "LessonPlanRequest"
      },
      "LessonPlanResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/LessonSection"
            },
            "type": "array",
            "title": "Sections"
          }
        },
        "type": "object",
        "required": [
          "title",
          "duration",
          "subject",
          "grade",
          "topic",
          "sections"
        ],
        "title": "LessonPlanResponse"
      },
...
      "PaperDigitizerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "originalImage": {
            "type": "string",
            "title": "Originalimage"
          },
          "digitizedText": {
            "type": "string",
            "title": "Digitizedtext"
          },
          "wordCount": {
            "type": "integer",
            "title": "Wordcount"
          },
          "format": {
            "type": "string",
            "title": "Format"
          }
        },
        "type": "object",
        "required": [
          "id",
          "originalImage",
          "digitizedText",
          "wordCount",
          "format"
        ],
        "title": "PaperDigitizerResponse"
      },
...
      "QuizGeneratorRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "questionCount": {
            "type": "integer",
            "title": "Questioncount"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "grade",
          "difficulty",
          "questionCount",
          "type"
        ],
        "title": "QuizGeneratorRequest"
      },
...
      "QuizResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestion"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "title",
          "subject",
          "grade",
          "questions"
        ],
        "title": "QuizResponse"
      },
...
      "RubricGeneratorRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "scale": {
            "type": "string",
            "title": "Scale"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "grade",
          "scale"
        ],
        "title": "RubricGeneratorRequest"
      },
...
      "RubricResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "levels": {
            "items": {
              "$ref": "#/components/schemas/RubricLevel"
            },
            "type": "array",
            "title": "Levels"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/RubricCriterion"
            },
            "type": "array",
            "title": "Criteria"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "levels",
          "criteria"
        ],
        "title": "RubricResponse"
      },
      "SimulationGeneratorRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "complexity": {
            "type": "string",
            "title": "Complexity"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "grade",
          "complexity"
        ],
        "title": "SimulationGeneratorRequest"
      },
...
      "SimulationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "logic": {
            "type": "string",
            "title": "Logic"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/SimulationParam"
            },
            "type": "array",
            "title": "Parameters"
          },
          "visualData": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Visualdata",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "topic",
          "logic",
          "parameters"
        ],
        "title": "SimulationResponse"
      },
...
      "StoryGeneratorRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "genre": {
            "type": "string",
            "title": "Genre"
          },
          "length": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "grade",
          "genre"
        ],
        "title": "StoryGeneratorRequest"
      },
      "StoryResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "tone": {
            "type": "string",
            "title": "Tone"
          },
          "moral": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moral"
          },
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/StoryChapter"
            },
            "type": "array",
            "title": "Chapters"
          }
        },
        "type": "object",
        "required": [
          "title",
          "tone",
          "chapters"
        ],
        "title": "StoryResponse"
      },
...
      "VisualGeneratorRequest": {
        "properties": {
          "sourceMode": {
            "type": "string",
            "title": "Sourcemode"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "bookId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bookid"
          },
          "chapterIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapterids"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style"
          }
        },
        "type": "object",
        "required": [
          "sourceMode"
        ],
        "title": "VisualGeneratorRequest"
      }
    }
  }
}